+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -60,8 +60,6 @@ import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 import { PortfolioService } from './portfolio.service';
 import { UpdateHoldingTagsDto } from './update-holding-tags.dto';
 
-const Fuse = require('fuse.js');
-
 @Controller('portfolio')
 export class PortfolioController {
   public constructor(
@@ -418,25 +416,13 @@ export class PortfolioController {
       filterByTags
     });
 
-    const { holdings: holdingsMap } = await this.portfolioService.getDetails({
+    const holdings = await this.portfolioService.getHoldings({
       dateRange,
       filters,
       impersonationId,
+      query: filterBySearchQuery,
       userId: this.request.user.id
     });
-
-    let holdings = Object.values(holdingsMap);
-
-    if (filterBySearchQuery) {
-      const fuse = new Fuse(holdings, {
-        keys: ['isin', 'name', 'symbol'],
-        threshold: 0.3
-      });
-
-      holdings = fuse.search(filterBySearchQuery).map(({ item }) => {
-        return item;
-      });
-    }
 
     return { holdings };
   }
